+++ examples/src/examples.cpp
@@ -249,9 +249,9 @@ void parse_exception_example()
 	{
         jsoncons::json val = jsoncons::json::parse(s);
     } 
-	catch(const jsoncons::json_parse_exception& e) 
+	catch(const jsoncons::parse_exception& e) 
 	{
-        std::cout << "Caught json_parse_exception with category " << e.code().category().name() 
+        std::cout << "Caught parse_exception with category " << e.code().category().name() 
 			      << ", code " << e.code().value() 
                   << " and message " << e.what() << std::endl;
     }