+++ test_suite/src/error_recovery_tests.cpp
@@ -23,12 +23,12 @@ class my_parse_error_handler : public parse_error_handler
 {
 private:
     virtual void do_warning(std::error_code ec,
-                            const parsing_context& context) throw(json_parse_exception)
+                            const parsing_context& context) throw(parse_exception)
     {
     }
 
     virtual void do_error(std::error_code ec,
-                          const parsing_context& context) throw(json_parse_exception)
+                          const parsing_context& context) throw(parse_exception)
     {
         if (ec.category() == json_error_category())
         {